+++ CountlyStarRating.m
@@ -150,7 +150,7 @@ const float kCountlyStarRatingButtonSize = 40;
 
         if(self.sessionCount == sessionCountSoFar)
         {
-            COUNTLY_LOG(@"Asking for star-rating as session count reached specified limit %i ...", self.sessionCount);
+            COUNTLY_LOG(@"Asking for star-rating as session count reached specified limit %d ...", (int)self.sessionCount);
 
             [self showDialog:^(NSInteger rating){}];
